Gradle

+++ FileReader/FileReader/src/net/qpowei/filereader/mc/nbt/tags/EndTag.java
@@ -12,7 +12,7 @@ public class EndTag extends Tag<Byte>
 	
 	
 	public EndTag() {
-		super("", 0);
+		super("", (byte)0);
 	}
 
 	@Override
